feat: New features and Bug fixes
The application had some functions in the wrong places, in other words, they were in places that could not use all the possibilities. For this to be fixed the functions have been moved to the parent App component. Some state bugs have also been found.

Visually it also had improvements, since more screens and more components were added, besides an optimization of the existing components, a new organization of the files was also organized so that everything stays in place so that the files do not mix or cause a code pollution. finally all the changes were:

* Change and optimization of functions
* Change in the visual
* Creation of new components
* Change in the organization of project files
* Creating Style.js files for each component
* Improved search and bugs fixed